+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Added total and unread counters for methods with changing read state
+
 ## [0.0.18] - 2024-02-09
 
 ### Changed